add exeption func

+++ cheat_scripts.py
@@ -3,9 +3,7 @@ from datacenter.models import Mark
 from datacenter.models import Chastisement
 from datacenter.models import Lesson
 from datacenter.models import Commendation
-from datacenter.models import Teacher
 import random
-from django.core.exceptions import MultipleObjectsReturned, ObjectDoesNotExist
 
 
 COMMENDATION_TEXTS = ['Молодец!', 'Отлично!', 'Хорошо!',
@@ -32,59 +30,48 @@ COMMENDATION_TEXTS = ['Молодец!', 'Отлично!', 'Хорошо!',
 
 
 def fix_marks(schoolkid_name):
-    try:
-        schoolkid_name = Schoolkid.objects.get(
-            full_name__contains=schoolkid_name)
-        udate_marks = Mark.objects.filter(
+    schoolkid_name = handle_schoolkid_exceptions(schoolkid_name)
+    if schoolkid_name:
+        update_marks = Mark.objects.filter(
             schoolkid=schoolkid_name, points__lt=4).update(points=5)
-        return udate_marks
-    except MultipleObjectsReturned:
-        print(f"Найдено более одного ученика по имени {schoolkid_name}, "
-              f"Укажите ФИО")
-    except ObjectDoesNotExist:
-        print(f"Ученик с именем {schoolkid_name} не найден")
+        return update_marks
 
 
 def delete_chastisements(schoolkid_name):
-    try:
-        schoolkid_name = Schoolkid.objects.get(
-            full_name__contains=schoolkid_name)
+    schoolkid_name = handle_schoolkid_exceptions(schoolkid_name)
+    if schoolkid_name:
         chastisements = Chastisement.objects.filter(schoolkid=schoolkid_name)
         chastisements.delete()
-    except MultipleObjectsReturned:
-        print(f"Найдено более одного ученика по имени {schoolkid_name}, "
-              f"Укажите ФИО")
-    except ObjectDoesNotExist:
-        print(f"Ученик с именем {schoolkid_name} не найден")
 
 
 def create_commendation(schoolkid_name, subject_title):
-    try:
-        schoolkid = Schoolkid.objects.get(full_name__contains=schoolkid_name)
-        subject = Lesson.objects.filter(subject__title=subject_title).first()
-        if subject:
-            lesson = Lesson.objects.filter(
-                subject__title=subject_title,
-                year_of_study=schoolkid.year_of_study,
-                group_letter=schoolkid.group_letter)
-
+    schoolkid = handle_schoolkid_exceptions(schoolkid_name)
+    if schoolkid:
+        lesson = Lesson.objects.filter(
+            subject__title=subject_title,
+            year_of_study=schoolkid.year_of_study,
+            group_letter=schoolkid.group_letter)
+        if lesson.first():
             lesson_date = lesson.last().date
-            teacher_name = lesson.first().teacher.full_name
-            teacher = Teacher.objects.get(full_name=teacher_name)
-
+            teacher = lesson.first().teacher
             commendation = Commendation.objects.create(
                 text=random.choice(COMMENDATION_TEXTS),
                 created=lesson_date,
                 schoolkid=schoolkid,
                 subject=lesson.first().subject,
                 teacher=teacher)
-
             return commendation
         else:
             print(f"Предмет с названием '{subject_title}' не найден")
 
-    except MultipleObjectsReturned:
+
+def handle_schoolkid_exceptions(schoolkid_name):
+    try:
+        schoolkid_name = Schoolkid.objects.get(
+            full_name__contains=schoolkid_name)
+        return schoolkid_name
+    except Schoolkid.MultipleObjectsReturned:
         print(f"Найдено более одного ученика по имени {schoolkid_name}, "
               f"Укажите ФИО")
-    except ObjectDoesNotExist:
+    except Schoolkid.DoesNotExist:
         print(f"Ученик с именем {schoolkid_name} не найден")
